test(download): move the real-socket downloadWithNodeHttp test to the integration lane
Address Codex P2 review finding on PR #6189: real loopback socket I/O
belongs in a *.integration.test.ts file, not a plain *.test.ts file,
which scripts/test-ts.sh classifies as hermetic unit-lane and runs
in parallel. Moved the one remaining real-socket happy-path test to
test/utils/FileDownloaderNodeHttp.integration.test.ts; the mid-body-close
regression coverage already uses a deterministic fake response stream
and stays in the unit lane.
