+++ test/utils/FileDownloader.test.ts
@@ -1,6 +1,5 @@
 import { afterEach, describe, expect, test } from "bun:test";
 import fs from "node:fs/promises";
-import http from "node:http";
 import path from "node:path";
 import { Readable } from "node:stream";
 import { DefaultFileDownloader } from "../../src/utils/FileDownloader";
@@ -12,18 +11,6 @@ type PipeResponseToFile = {
 
 const asPipeResponseToFile = (downloader: DefaultFileDownloader): PipeResponseToFile =>
   downloader as unknown as PipeResponseToFile;
-
-type NodeHttpDownloader = {
-  downloadWithNodeHttp(
-    url: string,
-    destination: string,
-    redirectCount: number,
-    signal?: AbortSignal,
-  ): Promise<void>;
-};
-
-const asNodeHttpDownloader = (downloader: DefaultFileDownloader): NodeHttpDownloader =>
-  downloader as unknown as NodeHttpDownloader;
 
 /**
  * A minimal stand-in for `http.IncomingMessage` that lets a test control
@@ -156,45 +143,9 @@ describe("DefaultFileDownloader pipeResponseToFile", function () {
   }, 500);
 });
 
-describe("DefaultFileDownloader downloadWithNodeHttp (end to end, real socket)", function () {
-  let tempDir: string | null = null;
-
-  afterEach(async function () {
-    if (tempDir) {
-      await fs.rm(tempDir, { recursive: true, force: true });
-      tempDir = null;
-    }
-  });
-
-  test("resolves with the full file for a complete real HTTP response", async function () {
-    const payload = Buffer.from("complete download payload");
-    const server = http.createServer((_request, response) => {
-      response.writeHead(200, { "content-length": String(payload.length) });
-      response.end(payload);
-    });
-    await new Promise<void>((resolve) => server.listen(0, "127.0.0.1", resolve));
-    const address = server.address();
-    if (address === null || typeof address === "string") {
-      throw new Error("Expected local HTTP server to listen on a TCP address");
-    }
-    const url = `http://127.0.0.1:${address.port}/payload`;
-
-    tempDir = await makeScratchTempDir("node-http-complete-");
-    const destination = path.join(tempDir, "file.bin");
-    // Calls the private downloadWithNodeHttp directly so the test exercises
-    // the Node HTTP fallback path itself, bypassing the curl/wget tiers
-    // that download() would otherwise prefer on a host that has them.
-    const downloader = asNodeHttpDownloader(new DefaultFileDownloader());
-
-    try {
-      await downloader.downloadWithNodeHttp(url, destination, 0);
-    } finally {
-      await new Promise<void>((resolve) => server.close(() => resolve()));
-    }
-
-    expect(await fs.readFile(destination)).toEqual(payload);
-  }, 2000);
-});
+// A real-socket end-to-end test of downloadWithNodeHttp lives in
+// test/utils/FileDownloaderNodeHttp.integration.test.ts — real loopback I/O
+// belongs in the integration lane, not this hermetic unit-test file.
 
 const makeScratchTempDir = async (prefix: string): Promise<string> => {
   const scratchDir = path.join(process.cwd(), "scratch");
